refactor: Separate worker pool
