+++ src/build.ts
@@ -3,6 +3,7 @@ import { getJson, resolvePaths } from "../utils/fs.ts";
 import { compileScripts } from "../utils/compileScripts.ts";
 import { getComponents } from "./getComponents.ts";
 import { generateRoutes } from "./generateRoutes.ts";
+import { createWorkerPool } from "./createWorkerPool.ts";
 import type { BuildWorkerEvent, ProjectMeta } from "../types.ts";
 
 async function build(projectMeta: ProjectMeta, projectRoot: string) {
@@ -18,7 +19,6 @@ async function build(projectMeta: ProjectMeta, projectRoot: string) {
   projectMeta.paths = resolvePaths(projectRoot, projectMeta.paths);
 
   const projectPaths = projectMeta.paths;
-
   const startTime = performance.now();
   const components = await getComponents("./components");
   const outputDirectory = projectPaths.output;
@@ -102,67 +102,6 @@ function getAmountOfThreads(
   }
 
   return amountOfThreads;
-}
-
-type WorkerStatus = "created" | "processing" | "waiting";
-type WorkerWrapper = { status: WorkerStatus; worker: Worker };
-
-function createWorkerPool<E>(amount: number, onWorkDone: () => void) {
-  const onReady = (workerWrapper: WorkerWrapper) => {
-    workerWrapper.status = "waiting";
-
-    if (waitingTasks.length) {
-      const task = waitingTasks.pop();
-
-      workerWrapper.status = "processing";
-      workerWrapper.worker.postMessage(task);
-    } else if (workers.every(({ status }) => status !== "processing")) {
-      onWorkDone();
-    }
-  };
-  const waitingTasks: E[] = [];
-  const workers: WorkerWrapper[] = Array.from(
-    { length: amount },
-    () => createWorker(onReady),
-  );
-
-  return {
-    addTaskToEach: (task: E) => {
-      workers.forEach(({ worker }) => worker.postMessage(task));
-    },
-    addTaskToQueue: (task: E) => {
-      const freeWorker = workers.find(({ status }) => status !== "processing");
-
-      if (freeWorker) {
-        freeWorker.status = "processing";
-        freeWorker.worker.postMessage(task);
-      } else {
-        waitingTasks.push(task);
-      }
-    },
-    terminate: () => {
-      workers.forEach(({ worker }) => worker.terminate());
-    },
-  };
-}
-
-function createWorker(onReady: (WorkerWrapper: WorkerWrapper) => void) {
-  const ret: WorkerWrapper = {
-    status: "created",
-    worker: new Worker(
-      new URL("./buildWorker.ts", import.meta.url).href,
-      {
-        type: "module",
-        deno: {
-          namespace: true,
-          permissions: "inherit",
-        },
-      },
-    ),
-  };
-  ret.worker.onmessage = () => onReady(ret);
-
-  return ret;
 }
 
 async function writeScripts(scriptsPath: string, outputPath: string) {
